+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Release date `UNRELEASED`
 - Add `handler/set-response-header`.
 - Implement `types/as-listener-builder` using protocol.
 - Implement `types/as-websocket-listener` using protocol.
+- Implement `types/as-websocket-connection-callback` using protocol.
 
 ## `1.0.72-rc2`